Fix non-UI test by moving it to a different class, as it can't share the same class initialization code.
