+++ Python/Tests/ProfilingUITests/ProfilingUITests.csproj
@@ -69,6 +69,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ProfilingUITests.cs" />
     <Compile Include="ProfilingTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
